Adding ability to configure wifiChecker.sh
Default settings are the same as before.
But now with the ability to read configuration
from brewpi-script/settings/config.cfg

A new argument "dumpconfig" is added to dump the configuration
to stdout.

Configuration options are:

"wifiCheckerDev" To specify wifi device the script should use.
"wifiCheckerEnabled" If the script should run or not.
"wifiCheckerMaxFailures" Total number of times to try and contact the
router if first packet fails
"wifiCheckerMaxFailures" Time to wait between failed attempts contacting the router