Add controller (advanced)
Socket:

Before this change, when adding a controller (/devices/add/)
only "Useinetsocket" was checked. When unchecked "Socket name"
was shown, check "Useinetsocket" again and "Sockethost" and
"Socketport" is shown.

With this change, "Useinetsocket" is checked as before as default
and "Sockethost" and "Socketport" is shown.

When form is created it will present a random port to use in the form,
this port is picked up by brewpi.py via dbcfg so no real need to
remember it.