Permanent Fermentrack Logo and User Image Upload Ability
Part One: Added Fermentrack Banner to Static and placed a call to display image in the flat_ui_template.

Part Two: Ability for user to upload image to be placed in the media file. Pillow Distribution must be part of the site packages.

Uploading of images to media is confirmed. User can also chose to use an external image.

How media image is called on the flat_ui_template is not the final intent and currently investigating how to change the img src