Redirect User upload Image if none available
Used onerror event to redirect User Logo if not yet uploaded to a default image.

Made some tweaks to settings and urls in an attempt to solve the image url pattern not working on the flat_ui_template. (changes didnt break anything, so were kept)

tThe need for Pillow installation was added to the requirement texts.