Ability to delete file from media directory
flat_ui_template calls for  brewerylogo.png to be the Brewert Logo to be displayed.

User must upload an image named brewerylogo.png for the image to be displayed.

If a user would like to change their logo. They will need to delete their current image (this will delete the whole instance from the DB including the file in the Medie Directory) and upload a new image named brewerylogo.png.